Changed put request

+++ src/main/java/com/altindag/hobbyproject/domain/Person.java
@@ -23,7 +23,7 @@ public class Person{
     private Long id;
     private String name;
     private int age;
-    private int balance;
+    private Long balance;
 
     public void setName(String name) {
         this.name = name.toUpperCase().charAt(0) + name.substring(1);
